+++ query-planner-js/src/__tests__/buildPlan.interfaceObject.test.ts
@@ -509,3 +509,108 @@ it('handles @requires on concrete type of field provided by interface object', (
     }
   `);
 });
+
+it('handles @interfaceObject in nested entity', () => {
+  const subgraph1 = {
+    name: 'S1',
+    typeDefs: gql`
+      type I @interfaceObject @key(fields: "id") {
+        id: ID!
+        t: T
+      }
+
+      type T {
+        relatedIs: [I]
+      }
+    `
+  }
+
+  const subgraph2 = {
+    name: 'S2',
+    typeDefs: gql`
+      type Query {
+        i: I
+      }
+
+      interface I @key(fields: "id") {
+        id: ID!
+        a: Int
+      }
+
+      type A implements I @key(fields: "id") {
+        id: ID!
+        a: Int
+      }
+
+      type B implements I @key(fields: "id") {
+        id: ID!
+        a: Int
+      }
+    `
+  }
+
+  const [api, queryPlanner] = composeAndCreatePlanner(subgraph1, subgraph2);
+
+  const operation = operationFromDocument(api, gql`
+    {
+      i {
+        t {
+          relatedIs {
+            a
+          }
+        }
+      }
+    }
+  `);
+
+  const plan = queryPlanner.buildQueryPlan(operation);
+  expect(plan).toMatchInlineSnapshot(`
+    QueryPlan {
+      Sequence {
+        Fetch(service: "S2") {
+          {
+            i {
+              __typename
+              id
+            }
+          }
+        },
+        Flatten(path: "i") {
+          Fetch(service: "S1") {
+            {
+              ... on I {
+                __typename
+                id
+              }
+            } =>
+            {
+              ... on I {
+                t {
+                  relatedIs {
+                    __typename
+                    id
+                  }
+                }
+              }
+            }
+          },
+        },
+        Flatten(path: "i.t.relatedIs.@") {
+          Fetch(service: "S2") {
+            {
+              ... on I {
+                __typename
+                id
+              }
+            } =>
+            {
+              ... on I {
+                a
+              }
+            }
+          },
+        },
+      },
+    }
+  `);
+});