+++ gateway-js/CHANGELOG.md
@@ -4,7 +4,7 @@ This CHANGELOG pertains only to Apollo Federation packages in the 2.x range. The
 
 ## vNext
 
-- Fix fed1 schema upgraded when @external is on a type [PR #2343](https://github.com/apollographql/federation/pull/2343).
+- Fix unexpected composition error about `@shareable` field when `@external` is on a type in a fed1 schema (one without `@link`) [PR #2343](https://github.com/apollographql/federation/pull/2343).
 
 ## 2.3.0-beta.3
 - Rewrites gateway response post-processing to avoid `@interfaceObject` related issues [PR 2335](https://github.com/apollographql/federation/pull/2335).